Changes progress loaders to be ProgressViews
